feature/refactor-common-code Refactor common code and arranged folder by code to enhance readbility
