+++ src/common/dto/document/delete-document.dto.ts
