+++ src/common/dto/user/create-user.dto.ts
@@ -1,5 +1,5 @@
 import { IsString, IsEmail, IsNotEmpty } from 'class-validator';
-import { BaseDto } from './base.dto';
+import { BaseDto } from '../base.dto';
 
 export class CreateUserDto extends BaseDto{
     @IsString()
